+++ game/components/base/graphics/c_text.h
@@ -1,0 +1,34 @@
+//project verse, 2017-2021
+//by jose pazos perez
+//all rights reserved uwu
+
+#pragma once
+
+#include "dtypes.h"
+#include "stb_truetype_fresa.h"
+
+namespace Verse
+{
+    struct FontInfo {
+        str path;
+        std::vector<ui8> data;
+        
+        ui64 font_size;
+        stbtt_fontinfo info;
+    };
+}
+
+namespace Verse::Component
+{
+    struct Text {
+        FontInfo* font;
+        
+        std::vector<ui8> bitmap;
+        Vec2 bitmap_size;
+        
+        ui32 tex_id;
+        Rect2 transform;
+        
+        int line_height;
+    };
+}